Add files via upload
Data Science and machine learning project  

Gole - Main Gole behind building this project is to know what actually data science and machine learning projects look like. 
And this is the first project I learned about machine learning and data science and how I could use that in the project. 

What I learned - 
In this project, 80% of the work is about data cleaning, which was not that easy, but after reviewing the project, it feels and looks interesting. 
Also, I also learned that in this field I need to build strong business knowledge and other soft skills like presentation skills, and communication skills. 

Data cleaning learnings - 
1. In the data cleaning area, I learned to find if there are null values and how to deal with them. 
2. How to remove outliers using standard deviation and mean. 


Standard deviation - The standard deviation of a random variable, sample, statistical population, data set, or probability distribution is the square root of its variance.

Mean - What is a mean in math?
The mean (aka the arithmetic mean, different from the geometric mean) of a dataset is the sum of all values divided by the total number of values. It's the most commonly used measure of central tendency and is often referred to as the “average.”